+++ src/model/Mora.java
@@ -1,40 +1,44 @@
 package model;
 
 import java.io.Serializable;
-import javax.persistence.*;
-import java.math.BigDecimal;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
+import javax.persistence.Table;
 
-/**
- * The persistent class for the mora database table.
- * 
- */
 @Entity
-@Table(name="mora")
-@NamedQuery(name="Mora.findAll", query="SELECT m FROM Mora m")
+@Table(name = "mora")
+@NamedQuery(name = "Mora.findAll", query = "SELECT m FROM Mora m")
 public class Mora implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	@Column(unique=true, nullable=false)
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(unique = true, nullable = false)
 	private int id;
 
-	@Column(name="FECHA_MORA", nullable=false, length=10)
+	@Column(name = "FECHA_MORA", nullable = false, length = 10)
 	private String fechaMora;
 
-	@Column(name="FECHA_VENCIMIENTO", nullable=false, length=10)
+	@Column(name = "FECHA_VENCIMIENTO", nullable = false, length = 10)
 	private String fechaVencimiento;
 
-	@Column(nullable=false, precision=10, scale=2)
-	private BigDecimal importe;
+	@Column(nullable = false, precision = 10, scale = 2)
+	private double importe;
 
-	@Column(nullable=false, length=1)
+	@Column(nullable = false, length = 1)
 	private String status;
 
-	//bi-directional many-to-one association to Contrato
-	@ManyToOne(fetch=FetchType.LAZY)
-	@JoinColumn(name="CONTRATO_ID", nullable=false)
+	// bi-directional many-to-one association to Contrato
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "CONTRATO_ID", nullable = false)
 	private Contrato contrato;
 
 	public Mora() {
@@ -64,11 +68,11 @@ public class Mora implements Serializable {
 		this.fechaVencimiento = fechaVencimiento;
 	}
 
-	public BigDecimal getImporte() {
+	public double getImporte() {
 		return this.importe;
 	}
 
-	public void setImporte(BigDecimal importe) {
+	public void setImporte(double importe) {
 		this.importe = importe;
 	}
